refactor(desktop): extract useBootstrap (init + account switch) from App.tsx (F5.7)
Startup and account switching move into useBootstrap: runInit (wait for
the backend session, detect missing credentials, load config flags /
keymap / labels / theme, then the inbox), the credentials import/retry
flow, and switchAccount (reset per-account UI state then reload). Pure
setState orchestration, lifted verbatim; the ~33 cross-subsystem setters
arrive via deps.

App.tsx 2,734 -> 2,574; useBootstrap 224 (< 500). unit 61 / e2e 36 green
(startup/theme flow covers runInit), tsc + build clean.
